Rename ApiSecret file
