+++ validator_client/src/http_api/api_secret.rs
